+++ Devystri/Devystri/Pages/Admin/AddApplication.cshtml.cs
@@ -75,7 +75,7 @@ namespace Devystri.Pages.Admin
                     toEdit.AppLogoName = ImportTools.ImageName(Application.AppLogo, toEdit.AppLogoName, imageImport);
                     toEdit.Presentation2RessourceName = ImportTools.ImageName(Application.Presentation2Ressource, toEdit.Presentation2RessourceName, imageImport);
                     toEdit.Presentation3RessourceName = ImportTools.ImageName(Application.Presentation3Ressource, toEdit.Presentation3RessourceName, imageImport);
-                    toEdit.Stat = (int)Application.State;
+                    toEdit.Stat = Application.State;
 
                     dbContext.Applications.Update(toEdit);
                     var sections = dbContext.Sections.Where(item => item.ProjectId == AppId).ToList();
